pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["TAMIPAMI"]

[project]
name = "tamipami"
version = "0.8.0"
description = "TBD"
authors = [
    {name = "Adam R. Rivers", email =  "adam.rivers@usda.gov"}]
maintainers = [
    {name = "Adam R. Rivers", email =  "adam.rivers@usda.gov"},
    {name = "Lei Ma", email = "lei.ma@usda.gov"}]
license = {file = "LICENSE.txt"}
readme = "README.rst"
keywords = ["Amplicon", "sequencing", "CRISPR", "TAM", "PAM"]
requires-python = ">=3.5"

dependencies = [
    "argpase",
    "biopython",
    "pandas",
    "scikit-bio",
    "logomaker",
    "statsmodels",
    "scipy",
    "matplotlib",
    "seaborn"
]


classifiers = [
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.5",
    "Development Status :: 3 - Alpha"
]



[project.urls]
repository = "http://github.com/microlei/tamipami"

[project.optional-dependencies]
tests = [
"pytest"
]


[project.scripts]
tamipami="TAMIPAMI.main:main"